Read a eBird personal sightings file

Your personal eBird sightings can be downloaded as a CSV file through the eBird website. This function reads the eBird data file, cleans up the variable names, and performs some basic processing.

eb_sightings(file, countable = FALSE)

Arguments

file
character; name of your personal eBird sightings files. This file can be downloaded from the eBird website.

countable
logical; whether to only return countable species, i.e. those that are counted on your life list by eBird.

Value

A dplyr::tibble, with additional class eb_sightings, containing your personal sightings.

Details

Some "species" reported on ebird are not true species. For example, they may be domestic or not resolved to the species level (e.g. Empidonax sp. or Greater/Lesser Yellowlegs), in which case they do not count towards your life list. Alternatively, sightings may be reported below the species level (e.g. subspecies or recognizable forms), in which case they must be rolled up to species level prior to reporting. Addiontal fields from the ebird taxonomy (eb_taxonomy) are added to the sightings data set to resolve these cases. The category field can be used to identify different categories of "species". The report_as, species_common, and species_scientific fields identify the true species corresponding to taxa reported below species. these fields are all NA for taxa not resolvable to species.

Examples

system.file("extdata/MyEBirdData.csv", package = "auklet") %>%
  eb_sightings()

#> # A tibble: 7,048 x 29
#>    submission_id taxon_order category species_code
#>            <chr>       <dbl>    <chr>        <chr>
#>  1     S18674562         214  species      wfwduc1
#>  2     S18586095         215  species       bbwduc
#>  3     S12725047         237  species       gwfgoo
#>  4      S9656769         237  species       gwfgoo
#>  5      S7828071         237  species       gwfgoo
#>  6     S28965809         237  species       gwfgoo
#>  7      S7828071         251 domestic      gragoo1
#>  8      S7828048         254  species       snogoo
#>  9      S7828071         254  species       snogoo
#> 10     S28961871         254  species       snogoo
#> # ... with 7,038 more rows, and 25 more variables: name_common <chr>,
#> #   name_scientific <chr>, order <chr>, family <chr>, report_as <chr>,
#> #   species_common <chr>, species_scientific <chr>, count <chr>, country <chr>,
#> #   state_province <chr>, county <chr>, location <chr>, latitude <dbl>,
#> #   longitude <dbl>, date <date>, time <time>, protocol <chr>, duration <int>,
#> #   all_obs_reported <lgl>, distance_traveled <dbl>, area_covered <chr>,
#> #   number_of_observers <int>, breeding_code <chr>, species_comments <chr>,
#> #   checklist_comments <chr>